+++ src/core/layers/sightings.py
@@ -247,7 +247,7 @@ class SammoSightingsLayer(SammoLayer):
             "validated",
             "sightNum",
             "observer",
-            "effortGroup",
+            "_effortGroup",
         ]:
             idx = layer.fields().indexFromName(field)
             form_config = layer.editFormConfig()
@@ -267,6 +267,20 @@ class SammoSightingsLayer(SammoLayer):
         layer.setDefaultValueDefinition(idx, QgsDefaultValue("''"))
 
     def _init_conditional_style(self, layer: QgsVectorLayer) -> None:
+        # dateTime
+        idx = layer.fields().indexFromName("dateTime")
+        cfg = {
+            "allow_null": False,
+            "calendar_popup": False,
+            "display_format": (
+                "dd/MM/yyyy HH:mm:ss "  # last space needed to avoid timezone
+            ),
+            "field_format": "yyyy-MM-dd HH:mm:ss",
+            "field_iso_format": False,
+        }
+        setup = QgsEditorWidgetSetup("DateTime", cfg)
+        layer.setEditorWidgetSetup(idx, setup)
+
         # side, distance
         expr = "@value is NULL"
         style = QgsConditionalStyle(expr)