+++ src/core/session.py
@@ -268,17 +268,30 @@ class SammoSession:
                 else StatusCode(int(bool(layer.featureCount())))
             )
         )
-        effortGroup = layer.maximumValue(layer.fields().indexOf("effortGroup"))
+        effortGroup = layer.maximumValue(
+            layer.fields().indexOf("_effortGroup")
+        )
         shipName = ""
         if self.surveyLayer.featureCount() > 0:
             shipName = next(self.surveyLayer.getFeatures())["shipName"]
         if effortGroup and statusCode == StatusCode.display(StatusCode.BEGIN):
             effortGroup += 1
+
+        effortLeg = 0
+        for ft in layer.getFeatures(
+            QgsFeatureRequest(QgsExpression(f"_effortGroup = {effortGroup}"))
+        ):
+            if ft["_effortLeg"] > effortLeg:
+                effortLeg = ft["_effortLeg"]
+        if statusCode != StatusCode.display(StatusCode.END):
+            effortLeg += 1
+
         self._addFeature(
             layer,
             geom=self.lastGpsInfo["geometry"],
             status=statusCode,
-            effortGroup=effortGroup or 1,
+            _effortGroup=effortGroup or 1,
+            _effortLeg=effortLeg or 1,
             speed=self.lastGpsInfo["gprmc"]["speed"],
             courseAverage=self.lastGpsInfo["gprmc"]["course"],
             shipName=shipName,
@@ -296,67 +309,96 @@ class SammoSession:
         # StatusCode.BEGIN to start a new route, and a new feature is created
         # with a StatusCode.END to end the previous route.
 
-        if (
-            not self.environmentLayer
-            or idx != self.environmentLayer.fields().indexOf("routeType")
-        ):
+        if not self.environmentLayer or idx not in [
+            self.environmentLayer.fields().indexOf("routeType"),
+            self.environmentLayer.fields().indexOf("status"),
+        ]:
             return
+        elif (
+            self.environmentLayer
+            and idx == self.environmentLayer.fields().indexOf("routeType")
+        ):
+            feat = self.environmentLayer.getFeature(fid)
+            request = QgsFeatureRequest().addOrderBy("dateTime", False)
+            for prevFeat in self.environmentLayer.getFeatures(request):
+                if prevFeat["fid"] == feat["fid"]:
+                    continue
+                elif prevFeat["status"] == StatusCode.display(
+                    StatusCode.END
+                ) or feat["status"] == StatusCode.display(StatusCode.END):
+                    return
+                elif (
+                    prevFeat["status"]
+                    in [
+                        StatusCode.display(StatusCode.BEGIN),
+                        StatusCode.display(StatusCode.ADD),
+                    ]
+                    and prevFeat["routeType"] == feat["routeType"]
+                ):
+                    self.environmentLayer.changeAttributeValue(
+                        fid,
+                        self.environmentLayer.fields().indexOf("status"),
+                        StatusCode.display(StatusCode.ADD),
+                    )
+                    self.environmentLayer.changeAttributeValue(
+                        fid,
+                        self.environmentLayer.fields().indexOf("_effortGroup"),
+                        prevFeat["_effortGroup"],
+                    )
+                elif prevFeat["routeType"] != feat["routeType"]:
+                    ft = QgsVectorLayerUtils.createFeature(
+                        self.environmentLayer
+                    )
+                    ft.setGeometry(feat.geometry())
+                    for attr in feat.fields().names():
+                        if attr in ["fid", "routeType", "dateTime", "status"]:
+                            continue
+                        ft[attr] = feat[attr]
+                    ft["routeType"] = prevFeat["routeType"]
+                    ft["dateTime"] = QDateTime(feat["dateTime"])
+                    ft["status"] = StatusCode.display(StatusCode.END)
+                    ft["_effortLeg"] = prevFeat["_effortLeg"]
+                    self.environmentLayer.addFeature(ft)
 
-        feat = self.environmentLayer.getFeature(fid)
-        request = QgsFeatureRequest().addOrderBy("dateTime", False)
-        for prevFeat in self.environmentLayer.getFeatures(request):
-            if prevFeat["fid"] == feat["fid"]:
-                continue
-            elif prevFeat["status"] == StatusCode.display(
-                StatusCode.END
-            ) or feat["status"] == StatusCode.display(StatusCode.END):
+                    self.environmentLayer.changeAttributeValue(
+                        fid,
+                        self.environmentLayer.fields().indexOf("status"),
+                        StatusCode.display(StatusCode.BEGIN),
+                    )
+                    self.environmentLayer.changeAttributeValue(
+                        fid,
+                        self.environmentLayer.fields().indexOf("_effortGroup"),
+                        prevFeat["_effortGroup"] + 1,
+                    )
+                    self.environmentLayer.changeAttributeValue(
+                        fid,
+                        self.environmentLayer.fields().indexOf("_effortLeg"),
+                        1,
+                    )
+                    self.environmentLayer.changeAttributeValue(
+                        fid,
+                        self.environmentLayer.fields().indexOf("dateTime"),
+                        ft["dateTime"].addSecs(1),
+                    )
+                break
+
+        elif (
+            self.environmentLayer
+            and idx == self.environmentLayer.fields().indexOf("status")
+        ):
+            feat = self.environmentLayer.getFeature(fid)
+            if feat["status"] != StatusCode.display(StatusCode.END):
                 return
-            elif (
-                prevFeat["status"]
-                in [
-                    StatusCode.display(StatusCode.BEGIN),
-                    StatusCode.display(StatusCode.ADD),
-                ]
-                and prevFeat["routeType"] == feat["routeType"]
-            ):
+            request = QgsFeatureRequest().addOrderBy("dateTime", False)
+            for prevFeat in self.environmentLayer.getFeatures(request):
+                if prevFeat["fid"] == feat["fid"]:
+                    continue
                 self.environmentLayer.changeAttributeValue(
                     fid,
-                    self.environmentLayer.fields().indexOf("status"),
-                    StatusCode.display(StatusCode.ADD),
+                    self.environmentLayer.fields().indexOf("_effortLeg"),
+                    prevFeat["_effortLeg"],
                 )
-                self.environmentLayer.changeAttributeValue(
-                    fid,
-                    self.environmentLayer.fields().indexOf("effortGroup"),
-                    prevFeat["effortGroup"],
-                )
-            elif prevFeat["routeType"] != feat["routeType"]:
-                ft = QgsVectorLayerUtils.createFeature(self.environmentLayer)
-                ft.setGeometry(feat.geometry())
-                for attr in feat.fields().names():
-                    if attr in ["fid", "routeType", "dateTime", "status"]:
-                        continue
-                    ft[attr] = feat[attr]
-                ft["routeType"] = prevFeat["routeType"]
-                ft["dateTime"] = QDateTime(feat["dateTime"])
-                ft["status"] = StatusCode.display(StatusCode.END)
-                self.environmentLayer.addFeature(ft)
-
-                self.environmentLayer.changeAttributeValue(
-                    fid,
-                    self.environmentLayer.fields().indexOf("status"),
-                    StatusCode.display(StatusCode.BEGIN),
-                )
-                self.environmentLayer.changeAttributeValue(
-                    fid,
-                    self.environmentLayer.fields().indexOf("effortGroup"),
-                    prevFeat["effortGroup"] + 1,
-                )
-                self.environmentLayer.changeAttributeValue(
-                    fid,
-                    self.environmentLayer.fields().indexOf("dateTime"),
-                    ft["dateTime"].addSecs(1),
-                )
-            break
+                break
 
     def addSightingsFeature(self) -> QgsVectorLayer:
         layer = self.sightingsLayer
@@ -364,10 +406,16 @@ class SammoSession:
         return layer
 
     def addFollowersFeature(
-        self, dt: str, geom: QgsGeometry, duplicate: bool
+        self, dt: str, geom: QgsGeometry, focalId: int, duplicate: bool
     ) -> None:
         layer = self.followersLayer
-        self._addFeature(layer, dt, geom, duplicate)
+        self._addFeature(
+            layer,
+            dt,
+            geom,
+            duplicate,
+            _focalId=focalId,
+        )
 
     def needsSaving(self) -> None:
         for layer in [
@@ -430,13 +478,13 @@ class SammoSession:
 
         # effort status check
         effortIds = environmentLayer.uniqueValues(
-            environmentLayer.fields().indexOf("effortGroup")
+            environmentLayer.fields().indexOf("_effortGroup")
         )
         errors = []
         for effortId in effortIds:
             effortIt = environmentLayer.getFeatures(
                 QgsFeatureRequest().setFilterExpression(
-                    f"effortGroup = {effortId}"
+                    f"_effortGroup = {effortId}"
                 )
             )
             statusCodes = {ft["datetime"]: ft["status"] for ft in effortIt}
@@ -511,40 +559,19 @@ class SammoSession:
         for feat in featuresIterator:
             if feat["validated"]:
                 continue
-
-            strDateTime = (
-                feat["dateTime"].toPyDateTime().strftime("%Y-%m-%d %H:%M:%S")
-            )
-            request = QgsFeatureRequest().setFilterExpression(
-                f"dateTime < to_datetime('{strDateTime}') "
-                f"and status != '{StatusCode.display(StatusCode.END)}'"
-            )
-            request.addOrderBy("dateTime", False)
-            for envFeat in environmentLayer.getFeatures(request):
-                if feat["side"] == "L":
+            if survey:
+                for attr in [
+                    "survey",
+                    "cycle",
+                    "computer",
+                ]:
+                    if attr == "computer" and feat["computer"]:
+                        continue
                     sightingsLayer.changeAttributeValue(
                         feat.id(),
-                        sightingsLayer.fields().indexOf("observer"),
-                        envFeat["left"],
+                        sightingsLayer.fields().indexOf(attr),
+                        survey[attr],
                     )
-                elif feat["side"] == "R":
-                    sightingsLayer.changeAttributeValue(
-                        feat.id(),
-                        sightingsLayer.fields().indexOf("observer"),
-                        envFeat["right"],
-                    )
-                elif feat["side"] == "C":
-                    sightingsLayer.changeAttributeValue(
-                        feat.id(),
-                        sightingsLayer.fields().indexOf("observer"),
-                        envFeat["center"],
-                    )
-                sightingsLayer.changeAttributeValue(
-                    feat.id(),
-                    sightingsLayer.fields().indexOf("effortGroup"),
-                    envFeat["effortGroup"],
-                )
-                break
 
             if survey:
                 sightingsLayer.changeAttributeValue(
@@ -577,26 +604,26 @@ class SammoSession:
         )
         idx = followersLayer.fields().indexOf("validated")
         for feat in featuresIterator:
+            if survey:
+                for attr in [
+                    "survey",
+                    "cycle",
+                    "session",
+                    "computer",
+                ]:
+                    if attr == "computer" and feat["computer"]:
+                        continue
+                    followersLayer.changeAttributeValue(
+                        feat.id(),
+                        followersLayer.fields().indexOf(attr),
+                        survey[attr],
+                    )
+
             followersLayer.changeAttributeValue(
                 feat.id(),
                 idx,
                 not merge,
             )
-            strDateTime = (
-                feat["dateTime"].toPyDateTime().strftime("%Y-%m-%d %H:%M:%S")
-            )
-            request = QgsFeatureRequest().setFilterExpression(
-                f"dateTime < to_datetime('{strDateTime}') "
-                f"and status != '{StatusCode.display(StatusCode.END)}'"
-            )
-            request.addOrderBy("dateTime", False)
-            for envFeat in environmentLayer.getFeatures(request):
-                followersLayer.changeAttributeValue(
-                    feat.id(),
-                    followersLayer.fields().indexOf("effortGroup"),
-                    envFeat["effortGroup"],
-                )
-                break
 
         followersLayer.commitChanges()
         followersLayer.startEditing()
@@ -640,7 +667,26 @@ class SammoSession:
         speed: Optional[float] = -9999.0,
         course: Optional[float] = -9999.0,
     ):
-        self._gpsLayer.add(longitude, latitude, hour, minu, sec, speed, course)
+        survey = ""
+        cycle = ""
+        computer = ""
+        if self.surveyLayer.featureCount() > 0:
+            ft = next(self.surveyLayer.getFeatures())
+            survey = ft["survey"]
+            cycle = ft["cycle"]
+            computer = ft["computer"]
+        self._gpsLayer.add(
+            longitude,
+            latitude,
+            hour,
+            minu,
+            sec,
+            speed,
+            course,
+            survey,
+            cycle,
+            computer,
+        )
 
     def _addFeature(
         self,
@@ -681,6 +727,9 @@ class SammoSession:
                         "strate",
                         "length",
                         "comment",
+                        "species",
+                        "podSize",
+                        "age",
                     ]:
                         continue
                     else: